refactor: extract useProviderSetup composable from Step3Provider
Split 698-line Step3Provider.vue into:
- useProviderSetup.ts (301 lines): provider CRUD, model import/delete,
  form state, error handling, dirty-reset watch
- Step3Provider.vue (458 lines): UI orchestration, transitions, step
  navigation, DEV helpers
